fix: 오류 수정

+++ src/main/java/com/ll/ticket/domain/admin/dto/RegisterConcertDto.java
@@ -1,6 +1,5 @@
 package com.ll.ticket.domain.admin.dto;
 
-import com.ll.ticket.domain.place.entity.Place;
 import com.ll.ticket.global.enums.ConcertCategory;
 import com.ll.ticket.global.enums.ConcertStatus;
 import jakarta.validation.constraints.Min;
@@ -9,7 +8,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.sql.Time;
 import java.time.LocalDateTime;
 
 @Getter
@@ -39,5 +37,5 @@ public class RegisterConcertDto {
 
     @NotNull(message = "좌석 가격을 입력하세요")
     @Min(0)
-    private int seatPrice;
+    private Long seatPrice;
 }
